backend/requirements.txt

fastapi==0.115.4
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.2
kafka-python==2.0.2
numpy==2.1.1
pandas==2.2.2
scikit-learn==1.5.2
pyarrow==17.0.0
itsdangerous==2.2.0
pydantic-settings>=2.2,<3
Mako>=1.3
Jinja2>=3.1
pgvector>=0.2.5
